Cleanup from PR review
Various bits of cleanup:
- Move and rename tests
- Use FunSpec to match codebase
- Use Koin to do injections
- Extract enum into own class
- Update strings to be more helpful
