+++ app/src/main/java/org/jellyfin/androidtv/constant/NetworkState.kt
@@ -1,0 +1,7 @@
+package org.jellyfin.androidtv.constant
+
+enum class NetworkState {
+	CONNECTED,
+	DISCONNECTED,
+	UNKNOWN
+}
