+++ app/src/main/java/org/jellyfin/androidtv/ui/global/NetworkStatusFragment.kt
@@ -6,10 +6,9 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import org.jellyfin.androidtv.R
-import org.jellyfin.androidtv.data.repository.NetworkState
+import org.jellyfin.androidtv.constant.NetworkState
 import org.jellyfin.androidtv.util.Utils
 import org.koin.android.ext.android.inject
