+++ app/src/main/java/org/jellyfin/androidtv/ui/global/NetworkStatusViewModel.kt
@@ -1,12 +1,10 @@
 package org.jellyfin.androidtv.ui.global
 
 import androidx.lifecycle.ViewModel
-import org.jellyfin.androidtv.data.repository.NetworkState
 import org.jellyfin.androidtv.data.repository.NetworkStatusRepository
 
 class NetworkStatusViewModel (
-	private val repo: NetworkStatusRepository
+	repo: NetworkStatusRepository
 ) : ViewModel() {
 	val getNetworkState = repo.state
-	fun getCurrentNetworkState() : NetworkState = repo.state.value
 }
